Changed send_DATA and send_RESP to use float to avoid problems with unchecked casts to int. Fixes a few issues Still not good repiratory cycle data

+++ C_Controller/ihm_communication.h
@@ -6,8 +6,8 @@
 // ------------------------------------------------------------------------------------------------
 //! Public interface to send event/data to the IHM and process read messages
 
-bool send_DATA(int P, int VolM, int Vol, int Pplat, int PEP);
-bool send_RESP(int IE, int FR, int VTe, int VM, int Pcrete, int Pplat, int PEP);
+bool send_DATA(float P, float VolM, float Vol, float Pplat, float PEP);
+bool send_RESP(float IE, float FR, float VTe, float VM, float Pcrete, float Pplat, float PEP);
 
 bool send_INIT(const char* information);
